Initial pack file
This is a 'build your own' UI kit.

The linked file is a Zip containing a PSD.

It's all vector-based, so you will likely need photoshop as that's what it was made in. Other applications may support this kit, but they are outside of the scope of technical help with this kit.

To use this kit, it is assumed that you can use photoshop, and can perform basic vector editing tasks. If not; There are many tutorials on youtube.

You will also need game assets to complete your UI, which cannot be provided here for copyright reasons. You can export them from the game yourself with a tool like <TOOLHERE>. Guidance for doing so can be found at the link.

You will also need two fonts:
Segoe UI (part of standard windows fonts)
Orbitron - https://www.fontsquirrel.com/fonts/orbitron

The kit is currently just a reproduction of standard in-game UI assets, and will be expanded over time.

Contents:
Single PSD in a Zip.
PSD contains the following elements:

Button frame
Text box
Bolt
Colour swatch
Font examples
6 Background panels

This pack is created and maintained by discord user Deantendo#4265 and can be contacted via the satisfactory modding channel's sub-channel; Art.
Contributions to this pack are welcome, though final acceptance and edits will be at the discretion of the pack maintainer for the time being or at such time as the pack is presented in a more open format.